+++ core/src/nucleus/ribosome/api/call.rs
@@ -1,24 +1,18 @@
 use crate::{
-    action::{Action, ActionWrapper},
-    context::Context,
     nucleus::{
-        is_fn_public,
-        reducers::execute_zome_function::launch_zome_fn_call,
+        actions::call_zome_function::call_zome_function,
         ribosome::{api::ZomeApiResult, Runtime},
-        state::NucleusState,
         ZomeFnCall,
     },
 };
 use holochain_core_types::{
-    dna::{capabilities::CapabilityCall, Dna},
-    entry::cap_entries::CapTokenGrant,
     error::HolochainError,
     json::JsonString,
 };
 use holochain_wasm_utils::api_serialization::{ZomeFnCallArgs, THIS_INSTANCE};
 use jsonrpc_lite::JsonRpc;
 use snowflake::ProcessUniqueId;
-use std::{convert::TryFrom, sync::Arc, time::Duration};
+use std::{convert::TryFrom};
 use wasmi::{RuntimeArgs, RuntimeValue};
 
 // ZomeFnCallArgs to ZomeFnCall
@@ -70,25 +64,8 @@ pub fn invoke_call(runtime: &mut Runtime, args: &RuntimeArgs) -> ZomeApiResult {
 fn local_call(runtime: &mut Runtime, input: ZomeFnCallArgs) -> Result<JsonString, HolochainError> {
     // ZomeFnCallArgs to ZomeFnCall
     let zome_call = ZomeFnCall::from_args(input);
-    // Create Call Action
-    let action_wrapper = ActionWrapper::new(Action::Call(zome_call.clone()));
-
-    let tick_rx = runtime.context.create_observer();
-    crate::instance::dispatch_action(runtime.context.action_channel(), action_wrapper);
-
-    loop {
-        if let Some(result) = runtime
-            .context
-            .state()
-            .unwrap()
-            .nucleus()
-            .zome_call_result(&zome_call)
-        {
-            return result;
-        } else {
-            let _ = tick_rx.recv_timeout(Duration::from_millis(10));
-        }
-    }
+    let context = &runtime.context;
+    context.block_on(call_zome_function(zome_call, context))
 }
 
 fn bridge_call(runtime: &mut Runtime, input: ZomeFnCallArgs) -> Result<JsonString, HolochainError> {
@@ -133,90 +110,6 @@ fn bridge_call(runtime: &mut Runtime, input: ZomeFnCallArgs) -> Result<JsonStrin
     }
 }
 
-pub fn validate_call(
-    context: Arc<Context>,
-    state: &NucleusState,
-    fn_call: &ZomeFnCall,
-) -> Result<Dna, HolochainError> {
-    if state.dna.is_none() {
-        return Err(HolochainError::DnaMissing);
-    }
-    let dna = state.dna.clone().unwrap();
-    // make sure the zome and function exists
-    let _ = dna
-        .get_function_with_zome_name(&fn_call.zome_name, &fn_call.fn_name)
-        .map_err(|e| HolochainError::Dna(e))?;
-
-    let public = is_fn_public(&dna, &fn_call)?;
-    if !public && !check_capability(context.clone(), &fn_call.clone()) {
-        return Err(HolochainError::CapabilityCheckFailed);
-    }
-    Ok(dna)
-}
-
-/// Reduce Call Action
-///   1. Checks for validity of ZomeFnCall
-///   2. Execute the exposed Zome function in a separate thread
-/// Send the result in a ReturnZomeFunctionResult Action on success or failure like ExecuteZomeFunction
-pub(crate) fn reduce_call(
-    context: Arc<Context>,
-    state: &mut NucleusState,
-    action_wrapper: &ActionWrapper,
-) {
-    // 1.Checks for correctness of ZomeFnCall
-    let fn_call = match action_wrapper.action().clone() {
-        Action::Call(call) => call,
-        _ => unreachable!(),
-    };
-
-    // 1. Validate the call (a number of things could go wrong)
-    let dna = match validate_call(context.clone(), state, &fn_call) {
-        Err(err) => {
-            // Notify failure
-            state.zome_calls.insert(fn_call.clone(), Some(Err(err)));
-            return;
-        }
-        Ok(dna) => dna,
-    };
-
-    // 2. Get the exposed Zome function WASM and execute it in a separate thread
-    let maybe_code = dna.get_wasm_from_zome_name(fn_call.zome_name.clone());
-    let code =
-        maybe_code.expect("zome not found, Should have failed before when validating the call.");
-    state.zome_calls.insert(fn_call.clone(), None);
-    launch_zome_fn_call(context, fn_call, &code, state.dna.clone().unwrap().name);
-}
-
-// TODO: check the signature too
-fn is_token_the_agent(context: Arc<Context>, cap: &Option<CapabilityCall>) -> bool {
-    match cap {
-        None => false,
-        Some(call) => context.agent_id.key == call.cap_token.to_string(),
-    }
-}
-
-/// checks to see if a given function call is allowable according to the capabilities
-/// that have been registered to callers in the chain.
-pub fn check_capability(context: Arc<Context>, fn_call: &ZomeFnCall) -> bool {
-    // the agent can always do everything
-    if is_token_the_agent(context.clone(), &fn_call.cap) {
-        return true;
-    }
-
-    match fn_call.cap.clone() {
-        None => false,
-        Some(call) => {
-            let chain = &context.chain_storage;
-            let maybe_json = chain.read().unwrap().fetch(&call.cap_token).unwrap();
-            let grant = match maybe_json {
-                Some(content) => CapTokenGrant::try_from(content).unwrap(),
-                None => return false,
-            };
-            grant.verify(call.cap_token.clone(), call.caller, &call.signature)
-        }
-    }
-}
-
 #[cfg(test)]
 pub mod tests {
     use super::*;
@@ -224,12 +117,13 @@ pub mod tests {
     extern crate wabt;
 
     use crate::{
+        action::{Action, ActionWrapper},
         context::Context,
         instance::{tests::test_instance_and_context, Instance},
         nucleus::{
             ribosome::{
                 api::{
-                    call::{Action, ActionWrapper, ZomeFnCall},
+                    call::ZomeFnCall,
                     tests::{
                         test_function_name, test_parameters, test_zome_api_function_wasm,
                         test_zome_name,
@@ -249,7 +143,7 @@ pub mod tests {
             fn_declarations::FnDeclaration,
             Dna,
         },
-        entry::Entry,
+        entry::{Entry, cap_entries::CapTokenGrant},
         error::{DnaError, HolochainError},
         json::JsonString,
     };
@@ -259,11 +153,9 @@ pub mod tests {
     use std::{
         collections::BTreeMap,
         sync::{
-            mpsc::{sync_channel, RecvTimeoutError},
+            mpsc::RecvTimeoutError,
             Arc,
         },
-        thread,
-        time::Duration,
     };
     use test_utils::create_test_dna_with_defs;
 
@@ -323,34 +215,10 @@ pub mod tests {
             "test",
             "{}",
         );
-        let zome_call_action = ActionWrapper::new(Action::Call(zome_call.clone()));
 
-        let (_, rx_observer) = sync_channel(1);
-        test_setup.instance.process_action(
-            zome_call_action,
-            Vec::new(),
-            &rx_observer,
-            &test_setup.context,
-        );
-
-        while test_setup
-            .instance
-            .state()
-            .nucleus()
-            .zome_call_result(&zome_call)
-            .is_none()
-        {
-            thread::sleep(Duration::from_millis(10));
-        }
-
-        let action_result = Ok(test_setup
-            .instance
-            .state()
-            .nucleus()
-            .zome_call_result(&zome_call)
-            .unwrap());
-
-        assert_eq!(expected, action_result);
+        let context = &test_setup.context;
+        let result = context.block_on(call_zome_function(zome_call, context));
+        assert_eq!(expected, Ok(result));
     }
 
     #[test]
@@ -467,17 +335,5 @@ pub mod tests {
                 let someone_else = Address::from("somoeone_else");
                 test_reduce_call(&test_setup,&String::from(addr),someone_else, expected_failure.clone());
         */
-    }
-
-    #[test]
-    fn test_agent_as_token() {
-        let dna = test_utils::create_test_dna_with_wat("bad_zome", "test_cap", None);
-        let test_setup = setup_test(dna);
-        let agent_token = Address::from(test_setup.context.agent_id.key.clone());
-        let context = test_setup.context.clone();
-        let cap_call = CapabilityCall::new(agent_token, None);
-        assert!(is_token_the_agent(context.clone(), &Some(cap_call)));
-        let cap_call = CapabilityCall::new(Address::from(""), None);
-        assert!(!is_token_the_agent(context, &Some(cap_call)));
     }
 }
